ci: Add/Implement MacOS builds based on new CMake Script
With the new CMake script we can actually offer MacOS builds of the plugin, though just like the other two platforms we have to take care of some dependencies ahead of time. zlib and libzip need to be built, and cached in order to keep build times low. For now we target MacOS 10.15 or higher, though this may change in the future.

The cache is controlled by the secret CACHE_VERSION, which can be changed in order to forcefully clear the cache for new builds.