ci: Update Ubuntu to new CMake script
The new CMake script takes care of a bit more, but now requires us to build our own zlib and libzip versions for each platform. This is a bit more complex to handle, but should not be too much of an issue for now.

While we had to remove the cache for libobs, obsdeps and qt, we do cache zlib and libzip now which brings us to almost the same build time. Additionally this removes the Ubuntu 18.04 builds, which means that the minimum Ubuntu version is now 20.04.

The cache is controlled by the secret CACHE_VERSION, which can be changed in order to forcefully clear the cache for new builds.