ui: Add Updater (#15)
Adds an automatic and manual check for updates in order to always keep users up to date with the latest changes. If an update is detected and the user wants to install it, an automated attempt at closing OBS Studio is performed, and the URL for the download is opened in a browser.